requirements.txt
click==8.0.1
Flask=2.0.1
RPi.GPIO==0.7.0 
